Camera, Character scripts
Have added camera script and character scripts. The camera should be
following the player (needs to be more thoroughly tested definitely), am
working on loading in textures and normals with the mesh.load method but
haven't got it working yet so it's commented out.

You'll need to download http://glew.sourceforge.net/ for the project to
work. The additional libraries etc are all set up already in the
property sheet, but you will need to match the folder address (ie
C/:Temp/glew-1.13.0)