Improved movement, added ground and concurrent keypress events as well as faster event response
Added some movement stuff, it's all handled in the main at the moment, I
was a bit confused how player (which is the enemy?) should be handled.

I changed the W,A,S,D keys to be handled inside the main loop, improving
responsiveness and allowing multiple events at once.

Add fog effects and ground.

At the moment the A, and D keys turn the character, later I'll bind that
to the mouse instead and have A and D strafe.

It'd be better to set boolean values in the if W,A,S,D key if
statements, and have if statements internal to the frame constraint if
statement checking these are updating them accordingly.

At the moment I just made the movementspeed and turnspeed variables
SUPER small to compensate. But that speed will be different in a release
version