Updated loader to register normals (and uvs)
The loader now separates out normals and uvs. Will cause buggy behaviour
at the moment if trying to load an obj without uvs so stick with cube2.
Was working with lighting on my older version of the project but I'll
need to look into making it work with the current version. So basically
at the moment it doesn't look like I've changed anything :D