shooty time
Projectile list added, and the enemy has a shoot class that fires a
projectile in the direction of the player, however there is no mesh
loading (as there is no mesh to load at the moment) and thus no drawing

the projectile lists are stored in the character thus we can check enemy
projectiles on player and vice-a-versa. Collisions not added yet.

Sorry I couldn't do more, but I have to go - Buenas Suerte