Comments and testing
just some comments.  I looked at the error Lance D in the code. not sure
if it is still their but it works when i ran the program.
however looked up the Exception thrown (nvoglv32.dll) it mentions stuff
about glDrawArray are we using that cause i cant find it in the code,
but this is the most used reason as to why that error happens